Move Scorer to flow.scorer
