+++ pyproject.toml
@@ -75,7 +75,7 @@ litellm = ["litellm>=1.36.1"]
 llamaindex = ["llama-index>=0.10.35"]
 mistral0 = ["mistralai>=0.1.8,<1.0.0"]
 mistral1 = ["mistralai>=1.0.0"]
-scorers = ["Levenshtein>=0.26.0", "instructor>=1.5.2"]
+scorers = ["Levenshtein>=0.26.0", "litellm>=1.58.2"]
 scorers_tests = [
   "Levenshtein>=0.26.0",
   "litellm>=1.58.2",
