+++ tests/scorers/test_llm_integrations.py
@@ -1,4 +1,5 @@
 import os
+
 import pytest
 
 from weave.scorers.summarization_scorer import (
@@ -28,10 +29,9 @@ PROVIDERS = {
 
 # Generate test parameters by iterating over each provider and its associated models.
 test_params = [
-    (provider, model)
-    for provider, cfg in PROVIDERS.items()
-    for model in cfg["models"]
+    (provider, model) for provider, cfg in PROVIDERS.items() for model in cfg["models"]
 ]
+
 
 @pytest.mark.parametrize("provider,model", test_params, ids=lambda p: f"{p[0]}:{p[1]}")
 @pytest.mark.asyncio
@@ -42,7 +42,7 @@ async def test_summarization_scorer_evaluate_summary(provider, model):
     # Check for the corresponding API key; skip the test if it's not present.
     if not os.getenv(env_key):
         pytest.skip(f"API key for {provider} not found. Skipping test.")
-    
+
     # Use the model directly as the model_id
     model_id = model
 
@@ -53,5 +53,7 @@ async def test_summarization_scorer_evaluate_summary(provider, model):
     )
     input_text = "The wolf is lonely in the forest. He is not happy that the fox is not with him."
     summary_text = "Wolf is lonely and missing the fox."
-    result = await summarization_scorer.evaluate_summary(input=input_text, summary=summary_text)
+    result = await summarization_scorer.evaluate_summary(
+        input=input_text, summary=summary_text
+    )
     assert isinstance(result, SummarizationEvaluationResponse)
