+++ tests/trace/test_evaluations.py
@@ -9,7 +9,7 @@ from PIL import Image
 import weave
 from tests.trace.util import AnyIntMatcher, AnyStrMatcher
 from weave import Evaluation, Model
-from weave.scorers import Scorer
+from weave.flow.scorer import Scorer
 from weave.trace.refs import CallRef
 from weave.trace.weave_client import get_ref
 from weave.trace_server import trace_server_interface as tsi
