+++ weave/scorers/__init__.py
@@ -1,13 +1,5 @@
-from weave.scorers.base_scorer import (
-    Scorer,
-    _has_oldstyle_scorers,
-    _validate_scorer_signature,
-    auto_summarize,
-    get_scorer_attributes,
-)
 from weave.scorers.classification_scorer import (
     MultiTaskBinaryClassificationF1,
-    transpose,
 )
 from weave.scorers.guardrails import PromptInjectionLLMGuardrail
 from weave.scorers.hallucination_scorer import HallucinationFreeScorer
@@ -34,8 +26,6 @@ __all__ = [
     "ContextEntityRecallScorer",
     "ContextRelevancyScorer",
     "EmbeddingSimilarityScorer",
-    "get_scorer_attributes",
-    "_has_oldstyle_scorers",
     "HallucinationFreeScorer",
     "InstructorLLMScorer",
     "ValidJSONScorer",
@@ -45,10 +35,7 @@ __all__ = [
     "OpenAIModerationScorer",
     "PromptInjectionLLMGuardrail",
     "PydanticScorer",
-    "Scorer",
     "StringMatchScorer",
     "SummarizationScorer",
-    "transpose",
     "ValidXMLScorer",
-    "_validate_scorer_signature",
 ]
