+++ weave/scorers/llm_utils.py
@@ -16,4 +16,3 @@ MISTRAL_DEFAULT_EMBEDDING_MODEL = "mistral-embed"
 
 DEFAULT_MAX_TOKENS = 4096
 DEFAULT_TEMPERATURE = 0.7
-
