+++ weave/scorers/moderation_scorer.py
@@ -3,7 +3,6 @@ from typing import Any, TypedDict
 from litellm import amoderation
 
 import weave
-from weave.scorers.base_scorer import Scorer
 from weave.scorers.llm_utils import OPENAI_DEFAULT_MODERATION_MODEL
 
 
@@ -14,7 +13,7 @@ class OpenAIModerationScorerOutput(TypedDict):
     categories: dict[str, bool]
 
 
-class OpenAIModerationScorer(Scorer):
+class OpenAIModerationScorer(weave.Scorer):
     """Use OpenAI moderation API to check if the model output is safe.
 
     Args:
