+++ weave/scorers/similarity_scorer.py
@@ -5,7 +5,7 @@ from litellm import aembedding
 from pydantic import Field
 
 import weave
-from weave.scorers.base_scorer import Scorer
+from weave.flow.scorer import Scorer
 from weave.scorers.llm_utils import OPENAI_DEFAULT_EMBEDDING_MODEL
 
 
@@ -45,7 +45,9 @@ class EmbeddingSimilarityScorer(Scorer):
         embeddings = await aembedding(self.model_id, [output, target])
         return embeddings.data[0]["embedding"], embeddings.data[1]["embedding"]
 
-    def cosine_similarity(self, vec1: list[float], vec2: list[float]) -> EmbeddingSimilarityScorerOutput:
+    def cosine_similarity(
+        self, vec1: list[float], vec2: list[float]
+    ) -> EmbeddingSimilarityScorerOutput:
         """Compute the cosine similarity between two vectors."""
         arr1 = np.array(vec1)
         arr2 = np.array(vec2)
